+++ src/index.ts
@@ -2,6 +2,7 @@ export * from './config';
 export * from './createStoreRegistry';
 export * from './loadingState';
 export * from './useList';
+export * from './useRecord';
 export * from './useActions';
 export * from './usePagination';
 export * from './useCrudState';